made code refactor
